backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.32.3
python-dotenv==0.19.0
google-generativeai==0.3.0
google-api-python-client==2.118.0
beautifulsoup4==4.9.3
feedparser==6.0.10
werkzeug==2.0.1
httpx>=0.24.1 
